--- lib/presentation/signin_page.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-class SigninPage extends ConsumerWidget {
-  const SigninPage({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return const Placeholder();
-  }
-}
